reorg user inputs and doc classes
